// react-module/src/openapi/models/storage-area-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface StorageAreaExtendedDto
 */
export interface StorageAreaExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof StorageAreaExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof StorageAreaExtendedDto
     */
    code?: any;
    /**
     * 
     * @type {WarehouseExtendedDto}
     * @memberof StorageAreaExtendedDto
     */
    warehouse?: any;
}